Basic scaffolding; rebased
